+++ src/components/dataTable.js
@@ -183,16 +183,6 @@
       }
     }, [data, searchTerm]);
 
-    // results caching fix
-    useEffect(() => {
-      if (!autoLoadOnScroll) {
-        const dataResults = data && data.results;
-        if (results.length === 0 && dataResults && dataResults.length > 0) {
-          setResults(dataResults);
-        }
-      }
-    }, [results]);
-
     useEffect(() => {
       const handler = setTimeout(() => {
         setSearchTerm(search);
@@ -442,6 +432,29 @@
     }, [results]);
 
     useEffect(() => {
+      if (pagination === 'never') {
+        const dataResults = data && data.results;
+        const needsCacheFix =
+          results.length === 0 && dataResults && dataResults.length > 0;
+
+        const setExistingData = () => {
+          setResults(dataResults);
+          fetchingNextSet.current = false;
+        };
+
+        if (needsCacheFix && !autoLoadOnScroll) {
+          setExistingData();
+        }
+        if (needsCacheFix && autoLoadOnScroll && skip === 0) {
+          setExistingData();
+        }
+        if (needsCacheFix && autoLoadOnScroll && skip !== 0) {
+          setSkip(0);
+        }
+      }
+    }, [results]);
+
+    useEffect(() => {
       if (isDev) {
         if (pagination === 'never') {
           setShowPagination(false);
@@ -480,8 +493,12 @@
       if (showPagination) {
         amount += paginationRef.current.clientHeight;
       }
-      if (amount > 0) {
-        const style = { height: `calc(100% - ${amount}px)` };
+      let style;
+      if (amount > 0 || !hasToolbar) {
+        style = {
+          height: `calc(100% - ${amount}px)`,
+          borderRadius: `${hasToolbar ? '0rem' : '0.1875rem'}`,
+        };
         setStylesProps({ style });
       } else {
         setStylesProps(null);
@@ -620,7 +637,11 @@
           style.getColor(backgroundHeader),
           '!important',
         ],
-        '& th': {
+        '& div': {
+          borderBottom: `${isDev ? '0.0625rem solid #cccccc' : 0}`,
+        },
+        '& th, & div[role="columnheader"]': {
+          borderBottom: `${isDev ? 0 : '0.0625rem solid #cccccc!important'}`,
           backgroundColor: ({ options: { backgroundHeader } }) => [
             style.getColor(backgroundHeader),
             '!important',
@@ -640,6 +661,7 @@
         pointerEvents: isDev && 'none',
       },
       pagination: {
+        borderRadius: '0.1875rem',
         pointerEvents: isDev && 'none',
         backgroundColor: ({ options: { background } }) => [
           style.getColor(background),